* Decoupling HMMs from the main library functionality to allow for other representations to be added more easily. * The Services are now create by decorating functions using the ServiceManager class.   * Simplifies the service generation by not needing to maintain a dictionary with all the services.   * Simplifies the class structure by not calling weird lambda functions anymore. * removing `qsr_type` from the service request and response.   * Now part of the general kwargs for hmm   * Particle filter will not use this   * Allows to have a generic hmm that takes trans and emi matrix plus a lookup instead of a specific implementation.
