+++ qsr_prob_rep/scripts/example_hmm_client.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 from qsrrep_ros.ros_client import ROSClient
-from qsrrep_lib.rep_io import HMMRepRequestCreate, HMMRepRequestSample, HMMRepRequestLogLikelihood
-from qsrrep_lib.rep_lib import ProbRepLib
+from qsrrep_lib.rep_io_hmm import HMMRepRequestCreate, HMMRepRequestSample, HMMRepRequestLogLikelihood
+#from qsrrep_lib.rep_lib import ProbRepLib
 import os
 import rospy
 import json
@@ -27,7 +27,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(parents=[general])
     subparsers = parser.add_subparsers(dest='action')
     qtc_parse = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,add_help=False)
-    qtc_parse.add_argument('-qsr', '--qsr_type', help="choose qsr: %s" % ProbRepLib.hmm_types_available.keys(), type=str, required=True)
+    qtc_parse.add_argument('-qsr', '--qsr_type', help="choose qsr: %s", type=str, required=True) #TODO: fix availabel
 
     # Parsers for create function
     create_parse = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,add_help=False)
@@ -57,7 +57,7 @@ if __name__ == "__main__":
 
     if args.action == "create":
         qsr_seq = load_files(args.input)
-        q, d = r.call_service(
+        d = r.call_service(
             HMMRepRequestCreate(
                 qsr_seq=qsr_seq,
                 qsr_type=args.qsr_type
@@ -67,7 +67,7 @@ if __name__ == "__main__":
 
     elif args.action == "sample":
         with open(args.input, 'r') as f: hmm = f.read()
-        q, s = r.call_service(
+        s = r.call_service(
             HMMRepRequestSample(
                 qsr_type=args.qsr_type,
                 xml=hmm,
@@ -83,7 +83,7 @@ if __name__ == "__main__":
     elif args.action == "loglikelihood":
         with open(args.qsr_seq, 'r') as f: qsr_seq = json.load(f)
         with open(args.input, 'r') as f: hmm = f.read()
-        q, l = r.call_service(
+        l = r.call_service(
             HMMRepRequestLogLikelihood(
                 qsr_type=args.qsr_type,
                 xml=hmm,
