+++ qsr_prob_rep/src/qsrrep_lib/rep_abstractclass.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+from abc import ABCMeta
+
+
+class RepAbstractclass():
+    """Abstract class for representations functionalities"""
+    __metaclass__ = ABCMeta
+
+    namespace = ""
