+++ qsr_prob_rep/srv/QSRProbRep.srv
@@ -1,7 +1,5 @@
 #Header header
-string qsr_type
 string data
 ---
 #Header header
-string qsr_type
 string data
